+++ src/SDKs/Compute/Compute.Tests/VMScaleSetTests/VMScaleSetScenarioTests.cs
@@ -29,7 +29,6 @@ namespace Compute.Tests
         /// </summary>
         [Fact]
         [Trait("Name", "TestVMScaleSetScenarioOperations")]
-        [Trait("Failure", "New Unable Match Http")]
         public void TestVMScaleSetScenarioOperations()
         {
             using (MockContext context = MockContext.Start(this.GetType().FullName))
@@ -51,7 +50,7 @@ namespace Compute.Tests
         /// Delete VMScaleSet
         /// Delete RG
         /// </summary>
-        [Fact(Skip = "ReRecord due to CR change")]
+        [Fact]
         [Trait("Name", "TestVMScaleSetScenarioOperations_ManagedDisks")]
         public void TestVMScaleSetScenarioOperations_ManagedDisks_PirImage()
         {
@@ -64,14 +63,14 @@ namespace Compute.Tests
         /// <summary>
         /// To record this test case, you need to run it again zone supported regions like eastus2euap.
         /// </summary>
-        [Fact(Skip = "ReRecord due to CR change")]
+        [Fact]
         [Trait("Name", "TestVMScaleSetScenarioOperations_ManagedDisks_PirImage_SingleZone")]
         public void TestVMScaleSetScenarioOperations_ManagedDisks_PirImage_SingleZone()
         {
             string originalTestLocation = Environment.GetEnvironmentVariable("AZURE_VM_TEST_LOCATION");
             try
             {
-                Environment.SetEnvironmentVariable("AZURE_VM_TEST_LOCATION", "eastus2euap");
+                Environment.SetEnvironmentVariable("AZURE_VM_TEST_LOCATION", "centralus");
                 using (MockContext context = MockContext.Start(this.GetType().FullName))
                 {
                     TestScaleSetOperationsInternal(context, hasManagedDisks: true, useVmssExtension: false, zones: new List<string> { "1" });
@@ -86,14 +85,14 @@ namespace Compute.Tests
         /// <summary>
         /// To record this test case, you need to run it again zone supported regions like eastus2euap.
         /// </summary>
-        [Fact(Skip = "ReRecord due to CR change")]
+        [Fact]
         [Trait("Name", "TestVMScaleSetScenarioOperations_ManagedDisks_PirImage_Zones")]
         public void TestVMScaleSetScenarioOperations_ManagedDisks_PirImage_Zones()
         {
             string originalTestLocation = Environment.GetEnvironmentVariable("AZURE_VM_TEST_LOCATION");
             try
             {
-                Environment.SetEnvironmentVariable("AZURE_VM_TEST_LOCATION", "eastus2euap");
+                Environment.SetEnvironmentVariable("AZURE_VM_TEST_LOCATION", "centralus");
                 using (MockContext context = MockContext.Start(this.GetType().FullName))
                 {
                     TestScaleSetOperationsInternal(context, hasManagedDisks: true, useVmssExtension: false, zones: new List<string> { "1", "3" });