+++ src/SDKs/DataFactory/Management.DataFactory/generate.ps1
@@ -1,0 +1,1 @@
+Start-AutoRestCodeGeneration -ResourceProvider "datafactory/resource-manager" -AutoRestVersion "latest"