+++ src/SDKs/PostgreSQL/Management.PostgreSQL/generate.ps1
@@ -1,1 +1,1 @@
-powershell.exe -ExecutionPolicy Bypass -NoLogo -NonInteractive -NoProfile -File "$(split-path $SCRIPT:MyInvocation.MyCommand.Path -parent)\..\..\..\..\tools\generateTool.ps1" -ResourceProvider "postgresql/resource-manager" -PowershellInvoker  -AutoRestVersion "latest"
+Start-AutoRestCodeGeneration -ResourceProvider "postgresql/resource-manager" -AutoRestVersion "latest"