+++ src/SDKs/Search/Build-SearchPackages.ps1
@@ -1,3 +1,10 @@
 $scriptDir = Split-Path $MyInvocation.MyCommand.Path -Parent
-msbuild "$scriptDir\..\..\..\build.proj" /t:PublishNuget /p:Scope="SDKs\Search\Management" /p:NugetPackageName="Microsoft.Azure.Management.Search"
-msbuild "$scriptDir\..\..\..\build.proj" /t:PublishNuget /p:Scope="SDKs\Search\DataPlane" /p:NugetPackageName="Microsoft.Azure.Search"
+$packagesDir = "$scriptDir\..\..\..\binaries\packages"
+
+if (!(Test-Path -Path $packagesDir))
+{
+    mkdir $packagesDir
+}
+
+msbuild "$scriptDir\..\..\..\build.proj" /t:CreateNugetPackage /p:Scope="SDKs\Search\Management"
+msbuild "$scriptDir\..\..\..\build.proj" /t:CreateNugetPackage /p:Scope="SDKs\Search\DataPlane"